[refactor #192] 관련 파일 caregiver 도메인으로 이동
