+++ src/main/java/com/sj/Petory/domain/caregiver/service/CareGiverService.java
@@ -6,9 +6,9 @@ import com.sj.Petory.domain.friend.repository.FriendStatusRepository;
 import com.sj.Petory.domain.member.dto.MemberAdapter;
 import com.sj.Petory.domain.member.entity.Member;
 import com.sj.Petory.domain.member.repository.MemberRepository;
-import com.sj.Petory.domain.pet.entity.CareGiver;
+import com.sj.Petory.domain.caregiver.entity.CareGiver;
 import com.sj.Petory.domain.pet.entity.Pet;
-import com.sj.Petory.domain.pet.repository.CareGiverRepository;
+import com.sj.Petory.domain.caregiver.repository.CareGiverRepository;
 import com.sj.Petory.domain.pet.repository.PetRepository;
 import com.sj.Petory.exception.FriendException;
 import com.sj.Petory.exception.MemberException;
